:recycle: route name rollback except pure board data
