+++ app/common/modal/Modal.tsx
@@ -1,68 +1,46 @@
-import {useState} from "react";
+import React, { useEffect, useRef } from "react";
 
 export default function Modal() {
-  const [showModal, setShowModal] = useState(false);
-  return (
-    <>
-      <button
-        className="bg-pink-500 text-white active:bg-pink-600 font-bold uppercase text-sm px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150"
-        type="button"
-        onClick={() => setShowModal(true)}
-      >
-        Open regular modal
-      </button>
-      {showModal ? (
-        <>
-          <div
-            className="z-50 justify-center items-center flex overflow-x-hidden overflow-y-auto fixed inset-0 outline-none focus:outline-none"
+  // Use useRef with the correct type for the dialog element
+  const modalRef = useRef<HTMLDialogElement>(null);
 
-          >
-            <div className="relative w-auto my-6 mx-auto max-w-3xl" onClick={() => {alert('2kooong2')}}>
-              <div className="border-0 rounded-lg shadow-lg relative flex flex-col w-full bg-white outline-none focus:outline-none">
-                <div className="flex items-start justify-between p-5 border-b border-solid border-blueGray-200 rounded-t">
-                  <h3 className="text-3xl font-semibold">
-                    Modal Title
-                  </h3>
-                  <button
-                    className="p-1 ml-auto bg-transparent border-0 text-black opacity-5 float-right text-3xl leading-none font-semibold outline-none focus:outline-none"
-                    onClick={() => setShowModal(false)}
-                  >
-                    <span className="bg-transparent text-black opacity-5 h-6 w-6 text-2xl block outline-none focus:outline-none">
-                      ×
-                    </span>
-                  </button>
-                </div>
-                <div className="relative p-6 flex-auto">
-                  <p className="my-4 text-blueGray-500 text-lg leading-relaxed">
-                    I always felt like I could do anything. That’s the main
-                    thing people are controlled by! Thoughts- their perception
-                    of themselves! They're slowed down by their perception of
-                    themselves. If you're taught you can’t do anything, you
-                    won’t do anything. I was taught I could do everything.
-                  </p>
-                </div>
-                <div className="flex items-center justify-end p-6 border-t border-solid border-blueGray-200 rounded-b">
-                  <button
-                    className="text-red-500 background-transparent font-bold uppercase px-6 py-2 text-sm outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150"
-                    type="button"
-                    onClick={() => setShowModal(false)}
-                  >
-                    Close
-                  </button>
-                  <button
-                    className="bg-emerald-500 text-white active:bg-emerald-600 font-bold uppercase text-sm px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150"
-                    type="button"
-                    onClick={() => setShowModal(false)}
-                  >
-                    Save Changes
-                  </button>
-                </div>
-              </div>
-              </div>          
-            </div>
-          <div className="opacity-25 fixed inset-0 z-40 bg-black" onClick={() => setShowModal(false)}></div>
-        </>
-      ) : null}
-    </>
+  useEffect(() => {
+    const modal = modalRef.current;
+
+    if (modal) {
+      const handleClickOutside = (event: MouseEvent) => {
+        // Ensure that event.target is of type HTMLDialogElement
+        if (event.target === modal) {
+          modal.close();
+        }
+      };
+
+      modal.addEventListener("click", handleClickOutside);
+
+      // Cleanup event listener on component unmount
+      return () => {
+        modal.removeEventListener("click", handleClickOutside);
+      };
+    }
+  }, []);
+
+  return (
+    <div>
+      <button
+        className="btn"
+        onClick={() => modalRef.current?.showModal()}
+      >
+        open modal
+      </button>
+      <dialog ref={modalRef} id="my_modal_2" className="modal">
+        <div className="modal-box">
+          <h3 className="font-bold text-lg">Hello!</h3>
+          <p className="py-4">Press ESC key or click outside to close</p>
+        </div>
+        <form method="dialog" className="modal-backdrop">
+          <button>close</button>
+        </form>
+      </dialog>
+    </div>
   );
 }
