+++ app/routes/board.$postId_.destroy.tsx
