+++ app/routes/board._index/components/NewBoard.tsx
@@ -11,7 +11,7 @@ export default function NewBoard() {
   const loading = fetcher.state !== "idle";
 
   return (
-      <fetcher.Form method="post" action="/api/board/create">
+      <fetcher.Form method="post" action="/board/create">
         {loading ? "loading..." : ""}
         <div>
           <label htmlFor="title">제목</label>
