+++ app/routes/board._index/route.tsx
@@ -5,7 +5,6 @@ import NewBoard from "./components/NewBoard";
 import { useState } from "react";
 import Center from "@/common/components/atoms/Center";
 import Modal from "@/common/modal/Modal";
-import AlternativeModal from "@/common/modal/AlternativeModal";
 
 
 export const loader = async () => {
@@ -34,7 +33,7 @@ export default function BoardRoute() {
           Click to show Form
         </button>
       )}
-      <AlternativeModal/>
+      <Modal/>
       {result?.data!.map((board, idx) => (
         <BoardItem key={idx} {...(board as BoardItemProps)}></BoardItem>
       ))}
